requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
filelock>=3.12.0
